增加注册和登录方法，增加短信校验的逻辑，增加token拦截校验，去掉ums中的shiro相关验证

+++ common/src/main/java/com/pipi/common/service/CheckMsgServiceImpl.java
@@ -79,7 +79,7 @@ public class CheckMsgServiceImpl implements CheckMsgService {
                 checkMsgRepository.save(ck);
             }
             // 然后新增一条验证码
-            String code = String.valueOf((Math.random() * 9 + 1) * 100000);
+            String code = String.valueOf(Double.valueOf((Math.random() * 9 + 1) * 100000).intValue());
             CheckMsg cm = new CheckMsg(phone, code, bizType);
             res = checkMsgRepository.save(cm);
         }
